Unit Testing

+++ src/main/java/com/exaltTraining/taskManagerProject/services/TeamServiceImpl.java
@@ -46,20 +46,27 @@ public class TeamServiceImpl implements TeamService {
     public Boolean assignTeamLeader(int teamId, int UserId) {
         Optional<User> tempUser= userRepository.findById(UserId);
         Optional<Team> tempTeam= teamRepository.findById(teamId);
-        if(tempUser.isPresent() && tempTeam.isPresent()){
-            User user = tempUser.get();
-            Team team = tempTeam.get();
+        try{
+            if(tempUser.isPresent() && tempTeam.isPresent()){
+                User user = tempUser.get();
+                Team team = tempTeam.get();
 
-            team.setTeamLeader(user);
-            user.setRole(User.Role.TEAM_MANAGER);
+                if(team.getTeamLeader()==null && team.getDepartment() == user.getDepartment()){
+                    team.setTeamLeader(user);
+                    user.setRole(User.Role.TEAM_MANAGER);
 
-            teamRepository.save(team);
-            userRepository.save(user);
-            notificationRepository.save(new Notification("Team leader", "You are assigned as a team leader for "+ team.getName() + "team", false, team.getTeamLeader()));
+                    teamRepository.save(team);
+                    userRepository.save(user);
+                    notificationRepository.save(new Notification("Team leader", "You are assigned as a team leader for "+ team.getName() + "team", false, team.getTeamLeader()));
 
-            return true;
-
+                    return true;
+                }
+            }
         }
+        catch(Exception e){
+            return false;
+        }
+
         return false;
     }
 
@@ -74,14 +81,23 @@ public class TeamServiceImpl implements TeamService {
             Team team = tempTeam.get();
             //check for department matching and not assigned to a team
             if(user.getDepartment()!=null){
-                System.out.println("in null department");
+
+                System.out.println(user.getDepartment());
+                System.out.println(departmentId);
                 if(user.getDepartment().getId() == departmentId) {
+
                     if (user.getTeam() == null) {
                         user.setTeam(team);
                         userRepository.save(user);
-                        notificationRepository.save(new Notification("Task Member assignment", "You are assigned as "+ team.getName() + "member", false, user));
+                        try{
+                            notificationRepository.save(new Notification("Task Member assignment", "You are assigned as "+ team.getName() + "member", false, user));
 
-                        return "The user has been assigned to the team";
+                            return "The user has been assigned to the team";
+                        }
+                        catch(Exception e){
+                            return null;
+                        }
+
 
                     } else {
                         return "The user is already assigned to a team";
@@ -103,28 +119,38 @@ public class TeamServiceImpl implements TeamService {
     }
     //Get all teams
     @Override
-    public List<Team> getAllTeams() {
-        return teamRepository.findAll();
+    public List<Team> getAllTeams(Department department) {
+        try{
+            return department.getTeams();
+        }
+        catch(Exception e){
+            return null;
+        }
+
     }
 
     //To delete a team by the department manager
     //The function checks that the team is really assigned with the department manager department
     @Override
     public String deleteTeam(int teamId, User departmentManager) {
-        Team team = teamRepository.findById(teamId).get();
-        if(team.getDepartment()== departmentManager.getDepartment())
-        {
-            try{
-                teamRepository.deleteById(teamId);
-                return "Team has been deleted";
+        Optional <Team> tempTeam = teamRepository.findById(teamId);
+        if(tempTeam.isPresent()){
+            Team team = tempTeam.get();
+            if(team.getDepartment()== departmentManager.getDepartment())
+            {
+                try{
+                    teamRepository.deleteById(teamId);
+                    return "Team has been deleted";
+                }
+                catch(Exception e){
+                    return "Error while deleting team";
+                }
             }
-            catch(Exception e){
-                return "Error while deleting team";
+            else{
+                return "The team is not assigned with this department!";
             }
         }
-        else{
-            return "The team is not assigned with this department!";
-        }
+        return "The team not found!";
 
     }
